Commented a lot of the code.
Fixed restrictions based on driver distraction level and vehicle speed.

Reworked the timer to start when the route is started. Pauses when the user exits to main menu. Resumes when the route is started again.

All AGA-values are now saved in another class that wont reset until the application is restarted.